Renamed the level geometry module
